+++ tests/helpers/helpers.js
@@ -1,0 +1,12 @@
+export function getRandomString(length) {
+    var randomChars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
+    var result = '';
+    for ( var i = 0; i < length; i++ ) {
+        result += randomChars.charAt(Math.floor(Math.random() * randomChars.length));
+    }
+    return result;
+}
+
+export function randomRepoPath() {
+    return `tests/test_repo_${getRandomString(10)}`;
+}