+++ tests/test_utils/api_utils.test.js
@@ -1,0 +1,12 @@
+import {checkServerDown, getUserForToken} from "../../src/utils/api_utils";
+
+
+test("Check Server Down", async () => {
+    expect(await checkServerDown()).toBe(false);
+});
+
+
+test("getUserForToken", async () => {
+    const res = await getUserForToken("abc");
+    expect(res.isTokenValid).toBe(false);
+});